created some folders
