--- Podfile
@@ -1,17 +1,0 @@
-# Uncomment the next line to define a global platform for your project
-# platform :ios, '9.0'
-
-target 'ToDo' do
-  # Comment the next line if you don't want to use dynamic frameworks
-  use_frameworks!
-
-  # Pods for ToDo
-  pod 'SwiftLint'
-  platform :ios, '12.0'
-  target 'ToDoTests' do
-    inherit! :search_paths
-    # Pods for testing
-    pod 'SwiftLint'
-  end
-
-end
