+++ ToDo.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "4c70d8de049006c4eb387db02086eb79de34c12fd28210628ece9b182749b4f9",
+  "originHash" : "b8dc8fb59d17b8be964398e38c97675112c458ca127d7983bc810cad49136c1f",
   "pins" : [
     {
       "identity" : "cocoalumberjack?tab=readme-ov-file",
