+++ ToDo/Model/ToDoItem/Model/FileCache/ToDoItem + FileCache.swift
@@ -1,0 +1,21 @@
+//
+//  ToDoItem + FileCache.swift
+//  ToDo
+//
+//  Created by Александр Андреев on 10.07.2024.
+//
+
+import Foundation
+extension ToDoItemModel {
+        
+    //сохранение в файл
+    func saveInFile(fileName: String) {
+        fileCache.saveInFile(items: self.toDos, fileName: fileName)
+        
+    }
+    //загрузка данных из файла
+    func uploadFromFile(fileName: String) -> [TodoItem] {
+        fileCache.uploadFromFile(fileName: fileName)
+    }
+
+}
